soulsync-ai/requirements.txt
fastapi==0.136.0
uvicorn==0.46.0
groq>=0.13.0
pydantic==2.13.3
python-dotenv==1.0.1
psycopg2-binary==2.9.10
pandas==3.0.2
numpy==2.4.3
sentence-transformers==3.4.1
faiss-cpu==1.10.0
openai-whisper==20250625
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.26
pydantic[email]>=2.0.0
